Refactoring according to comments, test should pass
